backend/pyproject.toml
[project]
name = "Study_Space"
version = "0.1.0"
description = "Study Space Backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "bcrypt==4.3.0",
    "fastapi[standard]>=0.116.1",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "psycopg[binary]>=3.2.9",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.43",
    "tiktoken>=0.11.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
]
